+++ client/store/index.js
@@ -10,7 +10,7 @@ import nearbyPlaces from './nearbyPlaces'
 import nextExperiences from './nextExperiences'
 import selectedPlace from './selectedPlace'
 import currentLocation from './currentLocation'
-import checkUsername from './checkUsername'
+import checkUsernameMessage from './checkUsernameMessage'
 import user from './user'
 import userError from './userError'
 import currentExperience from './experience'
@@ -33,7 +33,7 @@ const reducer = combineReducers({
   user,
   userError,
   nextExperiences,
-  checkUsername,
+  checkUsernameMessage,
   currentExperience,
   experience,
   routeId,
@@ -62,7 +62,7 @@ export * from './currentLocation'
 export * from './user'
 export * from './userError'
 export * from './nextExperiences'
-export * from './checkUsername'
+export * from './checkUsernameMessage'
 export * from './experience'
 export * from './routeId'
 export * from './currentRoute'
